# Conflicts:
#	libs/sub_client/client_qt/mainwindow.cpp
#	libs/sub_server/ClientReceiver.cpp

+++ libs/sub_server/server_box2d/Car.h
@@ -32,10 +32,10 @@ public:
     void jump();
     void turbo();
     void check_y_pos();
-    int X();
-    int Y();
+    float X();
+    float Y();
     float angle();
-    int angleDeg();
+    float angleDeg();
     uint8_t getId();
 };
